Branch commit splits
Added data structures that support splitting a Perforce submit into 1 or more git commits, each for their own branch.

The code here extends the original concept in p4-fusion to remove the hard coded references to 'HEAD' to instead be a condition when no streams or branch specs are requested, and in that case all submits are put into a single git branch commit value.

In the case of streams or branch specs, each commit is turned into branch move operations or single branch commit.  When the rest of main.cc and the change_list.cc are updated to take advantage of this, they will need to perform the merge, then run a separate commit on the same files to ensure any possible edit-on-integrate is handled correctly.