Branch Relative Paths
When dealing with branches, the code needs to know the relative path within the branch that will be put into the Git repository.  This adds that knowledge into the branch model.

At the moment, the branch spec model doesn't have enough information to make this mapping.  This will probably need to be changed to replace the use of an arbitrary branch spec to instead be a simplified model that looks a lot like a stream path - //(base location)/(branch name)/shared path location.  In this way, the stream functionality will be nearly identical.