Git Merge
The tool now performs a merge between branches in git.  The branch history shows up right.

There's a bug somewhere that seems to cause commits to happen to multiple branches.  There's something either used wrong with the HEAD and references, or there's a bug in the names being passed around.  This will require more investigation.