Fix for merging
The issue with the last commit is fixed.  Turned out, after changing 'HEAD' to point to the new branch, the index also had to change to reflect what HEAD now pointed to.  That, plus a bug with incorrectly marking branch to itself as a merge.

Also, this adds notes about a method for testing out the migration to ensure equal file contents and list of files for every change across branches.