Bug Fix and More Testing
There was a small bug in the file_data file, where the 'move/delete' was incorrectly marking the file as an integration.  These are the *source* of a file move, not a target, and so this file shouldn't be considered an integration action.

Updated the validation remarks to point instead to a shell script that analyzes the Perforce source and the generated Git repository to see if they are identical.  The validation is much more robust than the previous version.  There's some additional remarks about how one changelist can generate multiple commits, and how the validation works to resolve that scenario.