Remove the changelist number from the beginning of the commit message (#8)
Since in the BIMx repo almost all changelist descriptions already start with the Jira issue key, I think if the resulting Git commits are further prefixed with the p4 changelist number, it ends up being too much visual clutter when one is looking at history. The changelist number is also present in the added suffix so we are not losing information.